feat: optimisation improvements
