+++ src/authentication/index.ts
@@ -1,3 +1,3 @@
-export { AuthenticationProps, default as Authentication } from './components/Authentication';
+export { default as Authentication, type AuthenticationProps } from './components/Authentication';
 export { useAuthentication } from './context';
 export * from './types';
