+++ src/index.ts
@@ -1,8 +1,6 @@
-import { Organization } from './store/organizations';
-
 export * from './authentication';
 export * from './request';
 export * from './routing';
+export type { Organization } from './store/organizations';
 export * from './ui';
 export type { Optional } from './utils';
-export type { Organization };
