+++ src/ui/Navigation/components/Drawer/Navigation/components/MenuItem/MenuItem.tsx
@@ -1,12 +1,13 @@
-import { ExpandCollapse, Icon, useTheme, useTypeColorToColorMatch } from '@orfium/ictinus';
+import { ExpandCollapse, Icon, Theme, useTypeColorToColorMatch } from '@orfium/ictinus';
 import { AcceptedIconNames } from '@orfium/ictinus/dist/components/Icon/types';
 import { BASE_SHADE } from '@orfium/ictinus/dist/theme/palette';
-import React from 'react';
-import { useLocation, useRouteMatch } from 'react-router-dom';
+import React, { useMemo } from 'react';
+import { useRouteMatch } from 'react-router-dom';
 import { Optional } from '../../../../../../../utils';
 import FlippableArrow from '../../../../../../FlippableArrow';
 import { MenuIcon, MenuItemText } from '../../../../../common.styles';
 import { MenuItem as MenuItemType } from '../../../../../types';
+import { useMenuItemMatch } from './hooks';
 import {
   ArrowContainer,
   Bullet,
@@ -15,17 +16,23 @@ import {
   MenuLink,
 } from './MenuItem.styles';
 
+export type MenuItemProps = {
+  theme: Theme;
+  item: MenuItemType | Optional<MenuItemType, 'iconName'>;
+  isSubMenu?: boolean;
+};
+
 function MenuItemContent(props: {
-  expanded: boolean;
+  theme: Theme;
   item: Optional<MenuItemType, 'iconName'>;
+  expanded?: boolean;
   isSubMenu?: boolean;
 }) {
-  const { item, expanded, isSubMenu = false } = props;
+  const { theme, item, expanded = false, isSubMenu = false } = props;
   const match = useRouteMatch(item.url);
-  const theme = useTheme();
   const { calculateColorBetweenColorAndType } = useTypeColorToColorMatch();
-  const { shade } = calculateColorBetweenColorAndType('', 'primary');
 
+  const { shade } = calculateColorBetweenColorAndType('', 'primary');
   const isCurrent = !!match;
   const hasSubMenus = item.children && item.children.length > 0;
   const color = isCurrent
@@ -55,90 +62,59 @@ function MenuItemContent(props: {
   );
 }
 
-export type MenuItemProps = {
-  /** Defines if the menu item is expanded */
-  expanded: boolean;
-  toggleMenuItem: (newUrl: string) => void;
-  item: MenuItemType;
-};
+export function ExpandableMenuItem(props: MenuItemProps) {
+  const { item, theme } = props;
 
-function MenuItem({ expanded, toggleMenuItem, item }: MenuItemProps) {
-  const theme = useTheme();
-  const match = useRouteMatch(item.url);
-  const { state } = useLocation<{
-    previous: {
-      pathname: string;
-      search: string;
-    };
-  } | null>();
+  const { expanded, setExpanded, match } = useMenuItemMatch(item);
 
-  const hasSubMenus = item.children && item.children.length > 0;
+  const childrenElements = useMemo(() => {
+    return item.children
+      ? item.children.map((item) => <MenuItem key={item.url} theme={theme} item={item} isSubMenu />)
+      : null;
+  }, [item.children, theme]);
 
   return (
-    <React.Fragment>
-      {hasSubMenus ? (
-        <ExpandCollapseWrapper theme={theme} matched={!!match}>
-          <ExpandCollapse
-            expanded={expanded}
-            onChange={() => toggleMenuItem(item.url)}
-            textAndControl={(handleClick) => {
-              return (
-                <MenuItemButton
-                  theme={theme}
-                  type={'button'}
-                  data-testid={item.url}
-                  onClick={handleClick}
-                >
-                  <MenuItemContent expanded={expanded} item={item} />
-                </MenuItemButton>
-              );
-            }}
-          >
-            {() => {
-              return (
-                <React.Fragment>
-                  {item.children
-                    ? item.children.map((subMenuItem) => (
-                        <MenuLink
-                          theme={theme}
-                          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                          // @ts-ignore
-                          exact
-                          to={{
-                            pathname: subMenuItem.url,
-                            state,
-                          }}
-                          data-testid={subMenuItem.url}
-                          key={subMenuItem.url}
-                          id={'submenu-item-link'}
-                        >
-                          <MenuItemContent expanded={expanded} item={subMenuItem} isSubMenu />
-                        </MenuLink>
-                      ))
-                    : null}
-                </React.Fragment>
-              );
-            }}
-          </ExpandCollapse>
-        </ExpandCollapseWrapper>
-      ) : (
-        <MenuLink
-          theme={theme}
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          exact
-          to={{
-            pathname: item.url,
-            state,
-          }}
-          data-testid={item.url}
-          key={item.url}
-        >
-          <MenuItemContent expanded={expanded} item={item} />
-        </MenuLink>
-      )}
-    </React.Fragment>
+    <ExpandCollapseWrapper theme={theme} matched={!!match}>
+      <ExpandCollapse
+        expanded={expanded}
+        onChange={() => setExpanded((state) => !state)}
+        textAndControl={(handleClick) => {
+          return (
+            <MenuItemButton
+              theme={theme}
+              type={'button'}
+              data-testid={item.url}
+              onClick={handleClick}
+            >
+              <MenuItemContent theme={theme} expanded={expanded} item={item} />
+            </MenuItemButton>
+          );
+        }}
+      >
+        {() => {
+          return childrenElements;
+        }}
+      </ExpandCollapse>
+    </ExpandCollapseWrapper>
   );
 }
 
-export default MenuItem;
+export function MenuItem({ item, theme, isSubMenu = false }: MenuItemProps) {
+  const { historyState } = useMenuItemMatch(item);
+
+  return (
+    <MenuLink
+      theme={theme}
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      exact
+      to={{
+        pathname: item.url,
+        state: historyState,
+      }}
+      data-testid={item.url}
+    >
+      <MenuItemContent theme={theme} item={item} isSubMenu={isSubMenu} />
+    </MenuLink>
+  );
+}
