+++ src/ui/Navigation/components/Drawer/Navigation/components/MenuItem/index.ts
@@ -1,2 +1,1 @@
 export * from './MenuItem';
-export { default } from './MenuItem';
