+++ src/ui/Navigation/components/GlobalNav/index.ts
@@ -1,0 +1,2 @@
+export * from './GlobalNav';
+export { default } from './GlobalNav';
