+++ src/ui/Navigation/index.ts
@@ -1,3 +1,2 @@
 export * from './Navigation';
-export { default } from './Navigation';
 export * from './types';
