+++ src/ui/TopBar/index.ts
@@ -1,2 +1,1 @@
 export * from './TopBar';
-export { default } from './TopBar';
