+++ src/ui/index.ts
@@ -1,5 +1,3 @@
 export * from './consts';
 export * from './Navigation';
-export { default as Navigation } from './Navigation';
 export * from './TopBar';
-export { default as TopBar } from './TopBar';
